Extract special case: Aged Brie
Brie is the next-simplest special case to yoink out of our Pyramid o'
Ifs. We increment the quality, decrement the sell-in date, and then
increment again if we're past the sell-in.

We've re-introduced a bit of duplication here because the "decrement the
sell-in" line has been copied. That's fine for now; once we've separated
all the cases for clarity we can see about removing duplication in other
ways, if appropriate.