Move decrement_sell_in to end of operation
There doesn't seem to be a particular rationale for doing some quality
updates before the sell-in changes, and some after. We can just adjust
our conditionals and move the sell-in change to the end of each block.

I would have moved it out of the case statement entirely, but for the
hand of ragnaros the sell-in date never changes, so I'm not keen on any
implication that all items experience the passage of time.

I could equally have moved it to the start, but I preferred having "<=
0" as the boundary condition for past-sell-in items, whereas if I'd
moved it to the start it would've been "< -1" which is, like... weird. I
can't put it any better than that.