Extract object: DailyUpdateStrategy
I'd like to have different objects handle the updates for the different
items. I'm going to use a strategy pattern for this; as a first step,
this extracts a single object that implements the inner loop from the
GildedRose#update_quality method.

Because the #update_quality method actually causes the passage of time
as well as updating the quality, I've given our strategy what I think is
a better-named method, since the goblin won't let us change the
GildedRose API.