Extract per-item strategy objects
In the previous commit I extracted a single strategy object that just
used the same case statement as the GildedRose object had; this now
splits it into individual strategy objects for specific items.

If the goblin let me modify the item class I'd almost certainly have
done that instead, but the goblin is as the goblin does.

In the single case of this file, this extraction doesn't really
accomplish much; we've really just made the case statement a bit harder
to read, if I'm honest.

This sort of refactoring really comes into its own though if the objects
instantiated by the case statement are passed on to other bits of a more
complex system; it can be useful to have these named concepts in place
because it makes it easier to implement new, polymorphically varying
behaviour in future.