Use DailyUpdateStrategy superclass instead of UnknownItem
With the extraction of update_sell_in for the strategy subclasses, it
became clear that most of them were identical (time passes normally for
all but one item).

It therefore seemed reasonable to have the superclass handle this, and
the subclasses override it. And given that, it seemed even reasonabler
to just have the superclass deal with the default case, and let the
item-specific subclasses override whatever they need.