Move extra Item methods to DailyUpdateStrategy
I implemented those methods on Item before I became aware of the goblin.

After a frank exchange of views and the loss of a small amount of blood
I have agreed to put Item back the way it was, and the goblin has
generously agreed to remove its knife from my kidney.

Happily we now have a very plausible place to put said methods.